IDO: DTS: up to kernel 6.5.r7
soc: soc@ revert soc: soc
sdhc_2 revert 5.15 kernel.

With some of these changes, after starting android, a message appears that the SD card needs to be formatted. The recovery does not load and if you clear the date via fastboot, the recovery cannot mount userdata. Until I figure it out, let it stay that way.
____
The interconnect is removed in the gpu and turned off in the defconfig